backend/requirements.txt
# Core Framework
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Security & Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
cryptography==42.0.1

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# AI & Agents
pyautogen==0.2.10
openai==1.10.0
anthropic==0.8.1

# Vector Database & RAG
chromadb==0.4.22
sentence-transformers==2.3.1
langchain==0.1.4
langchain-community==0.0.17

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Validation & Serialization
pydantic==2.5.3

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# WebSockets
websockets==12.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
ruff==0.1.14
black==23.12.1
mypy==1.8.0

# Utilities
python-dateutil==2.8.2
